# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for managing shortcuts.",
        "title": "Multy API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/StartTurn/start": {
            "post": {
                "description": "Démarre la détection des messages StartTurn",
                "tags": [
                    "Start Turn"
                ],
                "summary": "Démarrer la détection du StartTurn",
                "responses": {
                    "200": {
                        "description": "StartTurn detection started",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to start StartTurn detection",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/StartTurn/stop": {
            "post": {
                "description": "Arrête la détection des messages StartTurn",
                "tags": [
                    "Start Turn"
                ],
                "summary": "Arrêter la détection du StartTurn",
                "responses": {
                    "200": {
                        "description": "StartTurn detection stopped",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to stop StartTurn detection",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/focus/{keyword}": {
            "post": {
                "description": "Met en avant une fenêtre qui contient un mot-clé spécifique dans son titre",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Windows"
                ],
                "summary": "Met en avant une fenêtre spécifique",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mot-clé pour identifier la fenêtre",
                        "name": "keyword",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fenêtre mise en avant avec succès",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la mise en avant de la fenêtre",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shortcut/register/{key}/{windowName}": {
            "post": {
                "description": "Registers a hotkey to focus on a window",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shortcut"
                ],
                "summary": "Register a hotkey",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Key to register",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of the window to focus",
                        "name": "windowName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/shortcut/unregister": {
            "delete": {
                "description": "Unregisters a previously registered keyboard shortcut",
                "tags": [
                    "Shortcut"
                ],
                "summary": "Unregister an existing hotkey",
                "responses": {
                    "200": {
                        "description": "Raccourci désenregistré avec succès",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to unregister shortcut",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/wheelclick/start": {
            "post": {
                "description": "Listens for middle mouse clicks and triggers click simulation on Dofus windows.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WheelClick"
                ],
                "summary": "Start middle mouse click detection",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/wheelclick/stop": {
            "post": {
                "description": "Stops the detection of middle mouse clicks.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WheelClick"
                ],
                "summary": "Stop middle mouse click detection",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/windows": {
            "get": {
                "description": "Obtient la liste des fenêtres actuellement ouvertes sur le système",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Windows"
                ],
                "summary": "Retourne la liste des fenêtres ouvertes",
                "responses": {
                    "200": {
                        "description": "Liste des fenêtres ouvertes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la récupération des fenêtres",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}